TIDIJ-222 Refactored Firebase manipulation code into separate classes #2 (forgot to stage)
