+++ UDP/entities/udp_server.py
@@ -1,6 +1,6 @@
 import socket
 
-from .strings import Strings
+from UDP.entities.utilities.strings import Strings
 
 
 class UDPServer:
